+++ lib/focused_controller.rb
@@ -1,4 +1,3 @@
 require 'focused_controller/version'
-require 'focused_controller/route'
 require 'focused_controller/mixin'
 require 'focused_controller/route_mapper'