+++ test/app/config/environments/test.rb
@@ -8,7 +8,13 @@ App::Application.configure do
   config.cache_classes = true
 
   # Configure static asset server for tests with Cache-Control for performance
-  config.serve_static_assets = true
+  if config.respond_to?(:serve_static_files)
+    config.serve_static_files = true
+  else
+    # Remove when we drop Rails 4.1
+    config.serve_static_assets = true
+  end
+
   config.static_cache_control = "public, max-age=3600"
 
   # Show full error reports and disable caching