+++ test/app/test/isolated_test_helper.rb
@@ -1,6 +1,0 @@
-require 'bundler/setup'
-require 'active_support/test_case'
-require 'focused_controller/test_helper'
-
-APP_ROOT = File.expand_path('../../app', __FILE__)
-POSTS    = []