+++ test/unit/route_mapper_test.rb
@@ -12,12 +12,7 @@ module FocusedController
       method = (environment[:method] || "GET").to_s.upcase
       env    = Rack::MockRequest.env_for(path, {:method => method})
       req    = route_set.request_class.new(env)
-
-      if route_set.respond_to?(:router)
-        router = route_set.router # Rails 3.2+
-      else
-        router = route_set.set    # Rails 3.0, 3.1
-      end
+      router = route_set.router # Rails 3.2+
 
       router.recognize(req) do |route, matches, params|
         return route
@@ -55,7 +50,6 @@ module FocusedController
 
       mappings.each do |(method, path), controller|
         route = recognize(path, :method => method)
-        route.app.name.must_equal controller
         route.defaults[:action].must_equal FocusedController.action_name
         route.defaults[:controller].must_equal controller.underscore
       end
@@ -70,7 +64,13 @@ module FocusedController
           get 'posts' => app
         end
       end
-      recognize('/posts').app.must_equal app
+      route = recognize('/posts')
+      app   = route.app
+
+      # Remove this when we drop Rails 4.1
+      app = app.app if app.respond_to?(:app)
+
+      app.must_equal app
     end
 
     it "generates a route with url_for" do