+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next version (2.0.0) ##
+
+* Rails 3.X no longer supported
+* Ruby 1.9 no longer officially supported (it may still work)
+
 ## Version 1.2.1 ##
 
 * Fix gemspec